chore: clippy warnings

+++ crates/lang/src/tipo/pattern.rs
@@ -438,13 +438,13 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                         if with_spread {
                             // Using the spread operator when you have already provided variables for all of the
                             // record's fields throws an error
-                            if pattern_args.len() == field_map.arity as usize {
+                            if pattern_args.len() == field_map.arity {
                                 return Err(Error::UnnecessarySpreadOperator {
                                     location: Span {
                                         start: location.end - 3,
                                         end: location.end - 1,
                                     },
-                                    arity: field_map.arity as usize,
+                                    arity: field_map.arity,
                                 });
                             }
 
@@ -465,7 +465,7 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                                 .position(|a| a.label.is_some())
                                 .unwrap_or(pattern_args.len());
 
-                            while pattern_args.len() < field_map.arity as usize {
+                            while pattern_args.len() < field_map.arity {
                                 let new_call_arg = CallArg {
                                     value: Pattern::Discard {
                                         name: "_".to_string(),
